test(hnsw): keep the child-logger test in compress_test.go
